+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 - xComputer:
-  - Fix for 'directory service is busy' Error When Joining a Domain and Renaming
-    a Computer when JoinOU is specified - Fixes [Issue #221](https://github.com/PowerShell/ComputerManagementDsc/issues/221).
+  - Fix for 'directory service is busy' error when joining a domain and renaming
+    a computer when JoinOU is specified - Fixes [Issue #221](https://github.com/PowerShell/ComputerManagementDsc/issues/221).
 
 ## 6.4.0.0